Safari-609.1.11.1/LayoutTests/http/tests/adClickAttribution/conversion-disabled-in-ephemeral-session-expected.txt

Tests that stored attributions cannot be converted in an ephemeral session.



--------
Frame: '<!--frame1-->'
--------
Conversion not received - timed out.

Unconverted Ad Click Attributions:
WebCore::AdClickAttribution 1
Source: 127.0.0.1
Destination: localhost
Campaign ID: 3
No conversion data.
